Add new 'ChangeMachine' class
The ChangeMachine attempts to generate a CoinBag with the smallest
number of coins possible that sum up to a given value.

The denominations of coins available to the ChangeMachine can vary.

If it's not possible to create exact change, the ChangeMachine will
over-compensate by returning an additional of the smallest denomination
coin available.

The ChangeMachine also has a feature which can restrict the number of
coins dispensed to a minimum. For example, assuming denominations
available of $0.05 and $0.01, if you request change for $0.05 but
restrict it to a minimum of 2 coins, you will receive 5x $0.01 coins
instead of 1x $0.05 coin.

Usage:
  x = ChangeMachine.new([1,5,10,25])
  bag = x.change(31, 3) # get change for $0.31 with a minimum of 3 coins